Bug fixes/improvements

+++ cos301_capstone/lib/User_Profile/Desktop_View.dart
@@ -1,26 +1,92 @@
 // ignore_for_file: prefer_const_literals_to_create_immutables, prefer_const_constructors
 
+import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:cos301_capstone/Edit_Profile/Edit_Profile.dart';
 import 'package:cos301_capstone/Global_Variables.dart';
 import 'package:cos301_capstone/Navbar/Desktop_View.dart';
 import 'package:cos301_capstone/Pets/Pet_Profile.dart';
 import 'package:cos301_capstone/services/HomePage/home_page_service.dart';
+import 'package:cos301_capstone/services/Profile/profile_service.dart';
 import 'package:cos301_capstone/services/general/general_service.dart';
 import 'package:dotted_border/dotted_border.dart';
 import 'package:flutter/material.dart';
 import 'package:animations/animations.dart';
 
 class ProfileDesktop extends StatefulWidget {
-  const ProfileDesktop({Key? key}) : super(key: key);
+  const ProfileDesktop({super.key, required this.userId});
 
+  final String userId;
   @override
   State<ProfileDesktop> createState() => _ProfileDesktopState();
 }
 
 class _ProfileDesktopState extends State<ProfileDesktop> {
+
+  ProfileDetails localProfileDetails = ProfileDetails();
+
+  String formatDate(DateTime date) {
+    return "${date.day} ${getMonthAbbreviation(date.month)} ${date.year}";
+  }
+
   @override
   void initState() {
     super.initState();
+
+    Future<void> getProfileDetails() async {
+      if (widget.userId != profileDetails.userID) {
+        Map<String, dynamic>? tempDetails = await ProfileService().getUserDetails(widget.userId);
+
+        if (tempDetails != null && tempDetails['profileVisibility']) {
+          localProfileDetails.userID = widget.userId;
+          localProfileDetails.name = tempDetails['name'];
+          localProfileDetails.surname = tempDetails['surname'];
+          localProfileDetails.email = tempDetails['email'];
+          localProfileDetails.bio = tempDetails['bio'];
+          localProfileDetails.profilePicture = tempDetails['profilePictureUrl'];
+          localProfileDetails.location = tempDetails['location'];
+          localProfileDetails.phone = tempDetails['phoneDetails']['phoneNumber'];
+          localProfileDetails.isoCode = tempDetails['phoneDetails']['isoCode'];
+          localProfileDetails.dialCode = tempDetails['phoneDetails']['dialCode'];
+          localProfileDetails.birthdate = formatDate(tempDetails['birthDate'].toDate());
+          localProfileDetails.userType = tempDetails['userType'];
+
+          localProfileDetails.pets = await ProfileService().getUserPets(widget.userId);
+          List<DocumentReference> localPosts = await ProfileService().getUserPosts(widget.userId);
+
+          profileDetails.myPosts.clear();
+
+          for (var post in localPosts) {
+            DocumentSnapshot postSnapshot = await post.get();
+            Map<String, dynamic> postData = postSnapshot.data() as Map<String, dynamic>;
+            postData['PostId'] = postSnapshot.id;
+            localProfileDetails.myPosts.add(postData);
+            print("Post data: $postData");
+          }
+        } else {
+          localProfileDetails.userID = widget.userId;
+          localProfileDetails.name = tempDetails!['name'];
+          localProfileDetails.surname = tempDetails['surname'];
+          localProfileDetails.bio = tempDetails['bio'];
+          localProfileDetails.profilePicture = tempDetails['profilePictureUrl'];
+
+          // Stubbed data
+          localProfileDetails.email = "Private Profile";
+          localProfileDetails.phone = "Private Profile";
+          localProfileDetails.birthdate = "Private Profile";
+          localProfileDetails.location = "Private Profile";
+          localProfileDetails.userType = "Private Profile";
+        }
+
+        setState(() {
+          print("Profile details set successfully for: ${localProfileDetails.userID}");
+        });
+      } else {
+        localProfileDetails = profileDetails;
+      }
+    }
+
+    getProfileDetails();
+
     profileDetails.isEditing.addListener(() {
       setState(() {});
     });
@@ -33,6 +99,12 @@ class _ProfileDesktopState extends State<ProfileDesktop> {
         });
       });
     });
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    profileDetails.isEditing.removeListener(() {});
   }
 
   @override
@@ -56,16 +128,16 @@ class _ProfileDesktopState extends State<ProfileDesktop> {
                 children: [
                   Expanded(
                     flex: 1,
-                    child: AboutMeContainer(),
+                    child: AboutMeContainer(profileDetails: localProfileDetails),
                   ),
                   SizedBox(width: 20),
                   Expanded(
                     flex: 3,
                     child: Column(
                       children: [
-                        MyPetsContainer(),
+                        MyPetsContainer(profileDetails: localProfileDetails),
                         SizedBox(height: 20),
-                        PostsContainer(),
+                        PostsContainer(localProfileDetails: localProfileDetails),
                       ],
                     ),
                   )
@@ -80,7 +152,9 @@ class _ProfileDesktopState extends State<ProfileDesktop> {
 }
 
 class PostsContainer extends StatefulWidget {
-  const PostsContainer({super.key});
+  const PostsContainer({super.key, required this.localProfileDetails});
+
+  final ProfileDetails localProfileDetails;
 
   @override
   State<PostsContainer> createState() => _PostsContainerState();
@@ -110,7 +184,7 @@ class _PostsContainerState extends State<PostsContainer> {
                 ),
               ),
               Divider(),
-              ListOfPosts(),
+              ListOfPosts(profileDetails: widget.localProfileDetails),
             ],
           ),
         ),
@@ -120,7 +194,9 @@ class _PostsContainerState extends State<PostsContainer> {
 }
 
 class ListOfPosts extends StatefulWidget {
-  const ListOfPosts({super.key});
+  const ListOfPosts({super.key, required this.profileDetails});
+
+  final ProfileDetails profileDetails;
 
   @override
   State<ListOfPosts> createState() => _ListOfPostsState();
@@ -134,7 +210,7 @@ class _ListOfPostsState extends State<ListOfPosts> {
         crossAxisAlignment: CrossAxisAlignment.start,
         mainAxisAlignment: MainAxisAlignment.start,
         children: [
-          for (var post in profileDetails.myPosts) ...[
+          for (var post in widget.profileDetails.myPosts) ...[
             Padding(
               padding: const EdgeInsets.symmetric(vertical: 8.0),
               child: Row(
@@ -185,28 +261,30 @@ class _ListOfPostsState extends State<ListOfPosts> {
                       ],
                     ),
                   ),
-                  IconButton(
-                    icon: Icon(Icons.delete, color: themeSettings.primaryColor),
-                    onPressed: () async {
-                      print("PostId: ${post["PostId"]}");
+                  if (widget.profileDetails.email == profileDetails.email) ...[
+                    IconButton(
+                      icon: Icon(Icons.delete, color: themeSettings.primaryColor),
+                      onPressed: () async {
+                        print("PostId: ${post["PostId"]}");
 
-                      bool deleted = await HomePageService().deletePost(post["PostId"]);
+                        bool deleted = await HomePageService().deletePost(post["PostId"]);
 
-                      if (deleted) {
-                        print("Post deleted successfully.");
-                        setState(() {
-                          profileDetails.myPosts.remove(post);
-                        });
-                      } else {
-                        ScaffoldMessenger.of(context).showSnackBar(
-                          SnackBar(
-                            content: Text('Failed to delete post'),
-                            backgroundColor: Colors.red,
-                          ),
-                        );
-                      }
-                    },
-                  )
+                        if (deleted) {
+                          print("Post deleted successfully.");
+                          setState(() {
+                            profileDetails.myPosts.remove(post);
+                          });
+                        } else {
+                          ScaffoldMessenger.of(context).showSnackBar(
+                            SnackBar(
+                              content: Text('Failed to delete post'),
+                              backgroundColor: Colors.red,
+                            ),
+                          );
+                        }
+                      },
+                    )
+                  ],
                 ],
               ),
             ),
@@ -218,7 +296,9 @@ class _ListOfPostsState extends State<ListOfPosts> {
 }
 
 class AboutMeContainer extends StatefulWidget {
-  const AboutMeContainer({super.key});
+  const AboutMeContainer({super.key, required this.profileDetails});
+
+  final ProfileDetails profileDetails;
 
   @override
   State<AboutMeContainer> createState() => _AboutMeContainerState();
@@ -241,15 +321,15 @@ class _AboutMeContainerState extends State<AboutMeContainer> {
           children: [
             CircleAvatar(
               radius: 100,
-              backgroundImage: NetworkImage(profileDetails.profilePicture),
+              backgroundImage: NetworkImage(widget.profileDetails.profilePicture),
             ),
             SizedBox(width: 20),
             Column(
               mainAxisAlignment: MainAxisAlignment.start,
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
-                Text("${profileDetails.name} ${profileDetails.surname}", style: TextStyle(fontSize: subHeadingTextSize)),
-                Text(profileDetails.bio, style: TextStyle(fontSize: subBodyTextSize)),
+                Text("${widget.profileDetails.name} ${widget.profileDetails.surname}", style: TextStyle(fontSize: subHeadingTextSize)),
+                Text(widget.profileDetails.bio, style: TextStyle(fontSize: subBodyTextSize)),
                 SizedBox(height: 20),
                 Text("Profile Details", style: TextStyle(fontSize: bodyTextSize, color: themeSettings.primaryColor)),
                 Divider(),
@@ -266,7 +346,7 @@ class _AboutMeContainerState extends State<AboutMeContainer> {
                       SizedBox(width: 10),
                       Flexible(
                         child: Text(
-                          profileDetails.email,
+                          widget.profileDetails.email,
                           style: TextStyle(fontSize: subBodyTextSize),
                           overflow: TextOverflow.ellipsis,
                         ),
@@ -287,7 +367,7 @@ class _AboutMeContainerState extends State<AboutMeContainer> {
                       SizedBox(width: 10),
                       Flexible(
                         child: Text(
-                          profileDetails.phone,
+                          widget.profileDetails.phone,
                           style: TextStyle(fontSize: subBodyTextSize),
                           overflow: TextOverflow.ellipsis,
                         ),
@@ -308,7 +388,7 @@ class _AboutMeContainerState extends State<AboutMeContainer> {
                       SizedBox(width: 10),
                       Flexible(
                         child: Text(
-                          profileDetails.birthdate,
+                          widget.profileDetails.birthdate,
                           style: TextStyle(fontSize: subBodyTextSize),
                           overflow: TextOverflow.ellipsis,
                         ),
@@ -329,7 +409,7 @@ class _AboutMeContainerState extends State<AboutMeContainer> {
                       SizedBox(width: 10),
                       Flexible(
                         child: Text(
-                          profileDetails.location,
+                          widget.profileDetails.location,
                           style: TextStyle(fontSize: subBodyTextSize),
                           overflow: TextOverflow.ellipsis,
                         ),
@@ -397,7 +477,9 @@ class _AboutMeContainerState extends State<AboutMeContainer> {
 }
 
 class MyPetsContainer extends StatefulWidget {
-  const MyPetsContainer({super.key});
+  const MyPetsContainer({super.key, required this.profileDetails});
+
+  final ProfileDetails profileDetails;
 
   @override
   State<MyPetsContainer> createState() => _MyPetsContainerState();
@@ -428,18 +510,70 @@ class _MyPetsContainerState extends State<MyPetsContainer> {
                 ),
               ),
               Divider(),
-              for (var pet in profileDetails.pets) ...[
+              for (var pet in widget.profileDetails.pets) ...[
+                if (widget.profileDetails.userID == profileDetails.userID) ...[
+                  OpenContainer(
+                    transitionDuration: Duration(milliseconds: 300),
+                    closedBuilder: (context, action) {
+                      return MouseRegion(
+                        cursor: SystemMouseCursors.click,
+                        child: Container(
+                          padding: const EdgeInsets.symmetric(vertical: 8.0),
+                          child: PetProfileButton(
+                            petName: pet["name"],
+                            petBio: pet["bio"],
+                            petPicture: pet["pictureUrl"],
+                          ),
+                        ),
+                      );
+                    },
+                    closedColor: Colors.transparent,
+                    closedElevation: 0,
+                    openBuilder: (context, action) {
+                      return PetProfile(
+                        creatingNewPet: false,
+                        petName: pet["name"],
+                        petBio: pet["bio"],
+                        petBirthdate: pet["birthDate"],
+                        petProfilePicture: pet["pictureUrl"],
+                        petID: pet["petID"],
+                      );
+                    },
+                  ),
+                ] else ...[
+                  Container(
+                    padding: const EdgeInsets.symmetric(vertical: 8.0),
+                    child: PetProfileButton(
+                      petName: pet["name"],
+                      petBio: pet["bio"],
+                      petPicture: pet["pictureUrl"],
+                    ),
+                  ),
+                ]
+              ],
+              SizedBox(height: 20),
+              if (widget.profileDetails.email == profileDetails.email) ...[
                 OpenContainer(
                   transitionDuration: Duration(milliseconds: 300),
                   closedBuilder: (context, action) {
                     return MouseRegion(
                       cursor: SystemMouseCursors.click,
-                      child: Container(
-                        padding: const EdgeInsets.symmetric(vertical: 8.0),
-                        child: PetProfileButton(
-                          petName: pet["name"],
-                          petBio: pet["bio"],
-                          petPicture: pet["pictureUrl"],
+                      child: DottedBorder(
+                        padding: EdgeInsets.all(20),
+                        borderType: BorderType.RRect,
+                        radius: Radius.circular(100),
+                        color: themeSettings.textColor,
+                        strokeWidth: 0.5,
+                        dashPattern: [5, 5], // Modify the dash pattern to make the border more spread out
+                        child: Row(
+                          children: [
+                            Icon(Icons.add, color: themeSettings.primaryColor, size: 30),
+                            SizedBox(width: 10),
+                            Text(
+                              "Add a new pet to your family",
+                              style: TextStyle(color: themeSettings.textColor),
+                            ),
+                          ],
                         ),
                       ),
                     );
@@ -448,50 +582,11 @@ class _MyPetsContainerState extends State<MyPetsContainer> {
                   closedElevation: 0,
                   openBuilder: (context, action) {
                     return PetProfile(
-                      creatingNewPet: false,
-                      petName: pet["name"],
-                      petBio: pet["bio"],
-                      petBirthdate: pet["birthDate"],
-                      petProfilePicture: pet["pictureUrl"],
-                      petID: pet["petID"],
+                      creatingNewPet: true,
                     );
                   },
                 ),
               ],
-              SizedBox(height: 20),
-              OpenContainer(
-                transitionDuration: Duration(milliseconds: 300),
-                closedBuilder: (context, action) {
-                  return MouseRegion(
-                    cursor: SystemMouseCursors.click,
-                    child: DottedBorder(
-                      padding: EdgeInsets.all(20),
-                      borderType: BorderType.RRect,
-                      radius: Radius.circular(100),
-                      color: themeSettings.textColor,
-                      strokeWidth: 0.5,
-                      dashPattern: [5, 5], // Modify the dash pattern to make the border more spread out
-                      child: Row(
-                        children: [
-                          Icon(Icons.add, color: themeSettings.primaryColor, size: 30),
-                          SizedBox(width: 10),
-                          Text(
-                            "Add a new pet to your family",
-                            style: TextStyle(color: themeSettings.textColor),
-                          ),
-                        ],
-                      ),
-                    ),
-                  );
-                },
-                closedColor: Colors.transparent,
-                closedElevation: 0,
-                openBuilder: (context, action) {
-                  return PetProfile(
-                    creatingNewPet: true,
-                  );
-                },
-              ),
               SizedBox(height: 20),
             ],
           ),
